Consider code spans when escaping Markdown styling
Previously, Markdown styling, such as _ characters, in an issue title
would always be escaped. This led to unnecessary escaping when the
styling characters were used within a code span (text wrapped in
backticks).

This commit disables the escaping of Markdown styling for text within
a code span. It does not consider the case where the code span uses
multiple backticks as its opening and closing delimiters. Such usage
in an issue title is considered to be unlikely and the added
complexity of coping with such delimiters was not felt to be
justified.

Fixes gh-87